requirements.txt
aiohttp
astroid
async-timeout
attrs
certifi
chardet
cleverbotfree
colorama
dblpy
deeppyer
discord.py
idna
isort
lazy-object-proxy
mccabe
motor
multidict
oauthlib
Pillow
praw
prawcore
pylint
pymongo
PySocks
requests
requests-oauthlib
selenium
six
statcord.py
toml
tweepy
typing-extensions
update-checker
urllib3
websocket-client
websockets
wrapt
yarl
pygelbooru
yippi